modifications ex10.3 et 10.4

+++ src/Etudiant.java
@@ -33,8 +33,8 @@ public class Etudiant {
         return somme / nbNotes;
     }
 
-    public double quelleMoyenne() {
-        return moyenne;
+    public int quelleMoyenne() {
+        return (int) moyenne;
     }
 
     public void afficheUnEtudiant() {
